+++ runtime-kotlin/runtime-desktop/feature/skillbill/src/commonMain/kotlin/skillbill/desktop/feature/skillbill/di/SkillBillComponent.kt
@@ -1,5 +1,6 @@
 package skillbill.desktop.feature.skillbill.di
 
+import skillbill.desktop.core.common.browser.BrowserLauncher
 import skillbill.desktop.core.common.di.ScreenScope
 import skillbill.desktop.core.common.di.UserScope
 import skillbill.desktop.core.ui.di.ScreenComponent
@@ -11,6 +12,7 @@ import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 @ContributesSubcomponent(ScreenScope::class)
 @SingleIn(ScreenScope::class)
 interface SkillBillComponent : ScreenComponent {
+  val browserLauncher: BrowserLauncher
   val viewModel: SkillBillViewModel
 
   @ContributesSubcomponent.Factory(UserScope::class)